requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
Flask-Migrate==3.1.0
python-dateutil==2.8.2
Werkzeug==2.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
supabase==2.3.0
gunicorn==21.2.0
